Move RAPTOR into knowledge compile package
